change: move controller/awsconfig to controller/legacy
