+++ service/controller/legacy/controller.go
@@ -1,9 +1,6 @@
-package awsconfig
+package legacy
 
 import (
-	// If your operator watches a CRD import it here.
-	// "github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
-
 	promclient "github.com/coreos/prometheus-operator/pkg/client/versioned"
 	"github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
 	"github.com/giantswarm/k8sclient/v3/pkg/k8sclient"
@@ -17,11 +14,11 @@ import (
 )
 
 type ControllerConfig struct {
+	BaseDomain       string
 	K8sClient        k8sclient.Interface
 	Logger           micrologger.Logger
 	PrometheusClient promclient.Interface
-
-	BaseDomain string
+	Provider         string
 }
 
 type Controller struct {
@@ -33,7 +30,12 @@ func NewController(config ControllerConfig) (*Controller, error) {
 
 	var resources []resource.Interface
 	{
-		c := controllerresource.Config(config)
+		c := controllerresource.Config{
+			BaseDomain:       config.BaseDomain,
+			K8sClient:        config.K8sClient,
+			Logger:           config.Logger,
+			PrometheusClient: config.PrometheusClient,
+		}
 
 		resources, err = controllerresource.New(c)
 		if err != nil {
@@ -43,12 +45,27 @@ func NewController(config ControllerConfig) (*Controller, error) {
 
 	var operatorkitController *controller.Controller
 	{
+		var runtimeObject runtime.Object
+		{
+			switch config.Provider {
+			case "aws":
+				runtimeObject = new(v1alpha1.AWSConfig)
+			case "azure":
+				runtimeObject = new(v1alpha1.AzureConfig)
+			case "kvm":
+				runtimeObject = new(v1alpha1.KVMConfig)
+			default:
+				return nil, microerror.Maskf(invalidProviderError, "provider: %q", config.Provider)
+			}
+
+		}
+
 		c := controller.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
-			Name:      "awsconfig-controller",
+			Name:      "legacy-controller",
 			NewRuntimeObjectFunc: func() runtime.Object {
-				return new(v1alpha1.AWSConfig)
+				return runtimeObject
 			},
 			Resources: resources,
 		}
